Moved parsers of built-in subs
