+++ src/built_ins/kill.rs
@@ -1,0 +1,36 @@
+pub mod parser {
+    use crate::built_ins::BuiltInSub;
+    use crate::common::*;
+    use crate::parser::pc::*;
+    use crate::parser::pc_specific::*;
+    use crate::parser::*;
+
+    pub fn parse<R>() -> impl Parser<R, Output = Statement>
+    where
+        R: Reader<Item = char, Err = QError> + HasLocation + 'static,
+    {
+        keyword_p(Keyword::Close).map(|_| Statement::BuiltInSubCall(BuiltInSub::Close, vec![]))
+    }
+}
+
+pub mod linter {
+    use crate::common::{QError, QErrorNode, ToErrorEnvelopeNoPos};
+    use crate::linter::arg_validation::ArgValidation;
+    use crate::parser::ExpressionNode;
+
+    pub fn lint(args: &Vec<ExpressionNode>) -> Result<(), QErrorNode> {
+        Ok(())
+    }
+}
+
+pub mod interpreter {
+    use crate::common::QError;
+    use crate::interpreter::interpreter_trait::InterpreterTrait;
+
+    pub fn run<S: InterpreterTrait>(interpreter: &mut S) -> Result<(), QError> {
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {}
