+++ src/parser/test_utils.rs
@@ -196,7 +196,7 @@ macro_rules! assert_built_in_sub_call {
         match result {
             Statement::BuiltInSubCall(actual_name, actual_args) => {
                 assert_eq!(actual_name, $expected_name);
-                let actual_args_no_loc: Vec<crate::parser::types::Expression> = actual_args.into_iter().map(|x| x.strip_location()).collect();
+                let actual_args_no_loc: Vec<crate::parser::Expression> = actual_args.into_iter().map(|x| x.strip_location()).collect();
                 assert_eq!(actual_args_no_loc, vec![$($arg),+]);
             }
             _ => panic!("Expected built-in sub call {:?}", $expected_name)
